Fix for compiling with gcc 3.1 and segfault in PerlEmbeddedInterpreter.cc

+++ Hook.cc
@@ -21,7 +21,6 @@
 #include <iomanip>
 #include <algorithm>  // min(), max()
 
-
 HookStub::HookStub(int p, float c, int n, bool F, bool en, bool col, string nm, vector<string> g) 
     : priority(p), chance(c), shots(n), fallthrough(F), enabled(en), color(col), name(nm), groups(g) 
 {}
@@ -558,7 +557,7 @@ TriggerHookStub::TriggerHookStub(int p, float c, int n, bool F, bool en, bool co
 
 bool TriggerHookStub::operator() (string& data, savedmatch* sm) {
     bool retval;
-    char out[std::max(2*command.length(),1024UL)]; // FIXME buffer overflow
+    char out[std::max(2*(unsigned long)command.length(),1024UL)]; // FIXME buffer overflow
 
     //if(sm) report_err("TriggerHookStub::operator() called with existing savedmatch!");
